Revise for PRAW 4
A lot was changed this commit. PRAW 4 was released, and it has a completely
different interface. To continue to work with the necessary library, the script
now calls the library correctly.

Additionally, some large structural changes were put in place. Main was
stripped of its data analysis functions, and those have been placed in the new
file "data_rake.py". Additionally, the helper class "TabList" has been
moved to its own file to aid in this.